feat: category and recipes to separate component
