+++ src/components/FoodSection/FoodSection.jsx
@@ -1,5 +1,32 @@
+import Category from "components/Category/Category";
+import RecipesComponent from "components/RecipesComponent";
+import { useSearchParams } from "react-router-dom";
+
 const FoodSection = () => {
-  return <></>;
+  const [searchParams, setSearchParams] = useSearchParams();
+  const category = searchParams.get("category") || "";
+  const area = searchParams.get("area") || "";
+  const ingredient = searchParams.get("ingredient") || "";
+
+  const handleSelectCategory = category => {
+    setSearchParams({ category, area: "", ingredient: "" });
+  };
+
+  const handleFiltersChange = (name, value) => {
+    const newParams = { category, area, ingredient, [name]: value };
+    setSearchParams(newParams);
+  };
+
+  return category === "" ? (
+    <Category onSelectCategory={handleSelectCategory} />
+  ) : (
+    <RecipesComponent
+      category={category}
+      area={area}
+      ingredient={ingredient}
+      onFiltersChange={handleFiltersChange}
+    />
+  );
 };
 
 export default FoodSection;
